Structure Project for Open Source Contributions
- separated components into three files to make contributions easy
